Second Commit

+++ Assets/TurnInfo.cs
@@ -5,5 +5,5 @@ internal class TurnInfo
 {
     public bool shoot { get; set; }
     public int time { get; set; }
-    public float rotation{get;set;}
+    public int rotation{get;set;}
 }
